==> 기업가치/part2/04. 상관분석/01. 재무지표1/01. code/PBR/ITS/5. make_index.ipynb <==
import pandas as pd
import numpy as np
import os
from scipy.cluster import hierarchy as hc
import matplotlib
import matplotlib.pyplot as plt
import warnings
import seaborn as sns
from scipy import stats
import math
from sklearn.preprocessing import MinMaxScaler

%matplotlib inline
warnings.filterwarnings(action='ignore')

matplotlib.rcParams['font.family'] ='Malgun Gothic'
matplotlib.rcParams['axes.unicode_minus'] =False

# Load data

# dir_root = 'C:/Users/kkt71/OneDrive/바탕 화면/기업가치분석'
dir_root = 'C:/Users/10188/Desktop/기업가치/상관분석/기업가치분석_data_result'
dir_master = os.path.join(dir_root, '01. data', 'master')
dir_processed = os.path.join(dir_root, '01. data', 'processed')
dir_result = os.path.join(dir_root, '03. result')

y = 'PBR'
c1 = 'Accenture'
c2 = 'Infosys'
c3 = 'TCS'
c4 = 'Cognizant'

dir_result_all = os.path.join(dir_result, 'All_ITS', y)

df = pd.read_excel(os.path.join(dir_result_all, 'corr_y_x_over7.xlsx'))

outperform = [c1, c2, c3]
lowperform = [c4]

target_out = [s+'_show' for s in outperform]
target_out = list(set(target_out)&set(df.columns))

target_low = [s+'_show' for s in lowperform]
target_low = list(set(target_low)&set(df.columns))

target_col = [s+'_show' for s in outperform+lowperform]
target_col = list(set(target_col)&set(df.columns))

# Make index

df['Outperform_avg'] = df[target_out].mean(axis=1)
df['Gap(Out~Low)'] = np.abs(df['Outperform_avg']-df[target_low[0]])
df['Std(Out)'] = df[target_out].std(axis=1)

scaler = MinMaxScaler()
df['Gap normed+1'] = pd.DataFrame(scaler.fit_transform(df[['Gap(Out~Low)']])+1)[0]

scaler = MinMaxScaler()
df['Std normed+1'] = pd.DataFrame(scaler.fit_transform(df[['Std(Out)']])+1)[0]

df['index 1'] = df['Gap normed+1']/df['Std normed+1']

scaler = MinMaxScaler()
df['Std normed original'] = pd.DataFrame(scaler.fit_transform(df[['Std(Out)']]))[0]

df['Std normed 0.5'] = df['Std normed original']*0.5 + 0.5
df['1/Std normed'] = 1/df['Std normed 0.5']

df['Index 2'] = 0.6*df['Gap normed+1']+0.4*df['1/Std normed']

df[[s+'_marked' for s in outperform+lowperform]+[s+'_plot' for s in outperform+lowperform]] = np.nan

writer = pd.ExcelWriter(os.path.join(dir_result_all, 'correlation_index.xlsx'), engine='xlsxwriter')
df.to_excel(writer, index=False)
writer.save()
